Moved Proxies to -client, and shared code out of the server to -shared, where it belongs. Cleaned up the client so it can build correctly, though it doesn't do a lot yet.
Very basic injection going on in both the client and the server - client should mostly be working, but the server still has some quirks, such as not tracking the current user. A session scope should be used there, but for that we need access to set the user (or data to get the user) in the session. The 'right' way seems to indicate that a provider should exist for the user, and the ServletScope.SESSION will read from it, but I don't yet see how that instance can be provided, unless that provided can change based on the request...

Cleaned up the poms a bit to keep webapp stuff in client, and out of everything else. Source version and resource encoding both moved up to the -parent module.