Added some profiles to make dev work easier:  o  db-init - executes the Bootstrap script in the war dir, so if using a embedded db, the file will be generated there.  o  dev-build - simplifies the compiled javascript somewhat, leaving out many features, but speeding up compiles. Among other things, split points are removed by this, so dev-mode builds are not suitable for any kind of acceptance testing.
Changed how the bytecode enhancement seems to work - I might be messing with fire here, but GEP doesn't like to launch with an agent, and RF can't handle the enhance-as-you-need it classes, so the types must be enhanced ahead of time. Note that Eclipse likes to ignore the specially prepared jar and grab the un-enhanced bytecode, so at the very least some documentation will be needed to deal with this.