Cleaned up the RF type used in login, thus removing the need to know how to serialize/deserialize all RF types in the app. See https://github.com/niloc132/tvguide-sample-parent/wiki/RequestFactory-Instances
This may not be a permenant change, its just something to try out. The change, made by removing get/set methods to other types in TvViewerProxy, also required messing with the ResolverServiceLayer to get it to not blow up when it looks at turning server data into client data.

After the change was complete, I tried restructuring the app startup to let us use the ginjector from the start of the app. Since the code it provides is lazily made available, it should be broken nicely into split points. Also learned that creating an interface for AsyncProxy and decorating it with @Inject calls will result in injection occuring on the proxy instance, starting it to load as soon as it is injected. Good to know for later.